\name{FastOccupancy-package}
\alias{FastOccupancy-package}
\alias{FastOccupancy}
\docType{package}
\title{
\packageTitle{FastOccupancy}
}
\description{
\packageDescription{FastOccupancy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FastOccupancy}
\packageIndices{FastOccupancy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FastOccupancy}

Maintainer: \packageMaintainer{FastOccupancy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
