% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotBaseLineCaptureProb}
\alias{plotBaseLineCaptureProb}
\title{Estimates of detection probability}
\usage{
plotBaseLineCaptureProb(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A plot with of the estimates. The interval represents the 
95\% credible interval.
}
\description{
Plots the estimates of the detection probability. This will plot
one value if \code{usingYearDetProb} is set to \code{FALSE} and if set to \code{TRUE}
will plot as many values as the number of years.
}
\examples{

plotBaseLineCaptureProb(sampleResults)

}
