% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.r
\name{plotGOFSpaceDections}
\alias{plotGOFSpaceDections}
\title{Goodness of fit of the detection in each region}
\usage{
plotGOFSpaceDections(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A map with the spatial points used in the approximation. The color
used varies according to wether the true value of statistcs is inside, above 
or below the 95\% credible interval.
}
\description{
This function performs a goodness of fit of the detection
in each region as described in Diana et al. (2021).
}
\examples{

plotGOFSpaceDections(sampleResults)

}
