% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.r
\name{plotGOFYearDetections}
\alias{plotGOFYearDetections}
\title{Goodness of fit of the yearly detection}
\usage{
plotGOFYearDetections(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A plot with the goddness of fit. The red points represent the true counts and
the interval represents the posterior of the yearly counts as estimated from the model
}
\description{
This function performs a goodness of fit of the yearly detection
as described in Diana et al. (2021).
}
\examples{

plotGOFYearDetections(sampleResults)

}
