% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotOccupancyCovariatesEffect}
\alias{plotOccupancyCovariatesEffect}
\title{Estimates of covariates on occupancy probability}
\usage{
plotOccupancyCovariatesEffect(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A plot with of the estimates. The interval represents the 
95\% credible interval.
}
\description{
Plots the estimates of the covariates on the occupancy probability.
}
\examples{

plotOccupancyCovariatesEffect(sampleResults)

}
