% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotOccupancyIndex}
\alias{plotOccupancyIndex}
\title{Occupancy index}
\usage{
plotOccupancyIndex(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A plot with of the estimates. The interval represents the 
95\% credible interval and the dots represent the median.
}
\description{
Plots the estimates of the occupancy index, 
as described in Diana et al. (2021).
}
\examples{

plotOccupancyIndex(sampleResults)

}
