% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotSpatialSitesEffect}
\alias{plotSpatialSitesEffect}
\title{Spatial Site Effect}
\usage{
plotSpatialSitesEffect(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
The posterior means of the spatial effect of each site on the map.
}
\description{
Plots the estimates of the the spatial site effects
}
\examples{

plotSpatialSitesEffect(sampleResults)

}
