% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotYearlyAveragePresenceProb}
\alias{plotYearlyAveragePresenceProb}
\title{Yearly occupancy probabilities}
\usage{
plotYearlyAveragePresenceProb(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A plot with of the estimates. The interval represents the 
95\% credible interval.
}
\description{
Plots the estimates of the average yearly occupancy
probabilities across all sites, as described in Diana et al. (2021).
}
\examples{

plotYearlyAveragePresenceProb(sampleResults)

}
