% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printBaseLineCaptureProb}
\alias{printBaseLineCaptureProb}
\title{Estimates of detection probability}
\usage{
printBaseLineCaptureProb(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A data.frame with the 95\% credible interval and the median.
}
\description{
Prints the estimates of the detection probability.
}
\examples{

printBaseLineCaptureProb(sampleResults)

}
