% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printDetectionCovariatesEffect}
\alias{printDetectionCovariatesEffect}
\title{Estimates of covariates on detection probability}
\usage{
printDetectionCovariatesEffect(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A data.frame with the 95\% credible interval and the median.
}
\description{
Prints the estimates of the covariates on the detection probability.
}
\examples{

printDetectionCovariatesEffect(sampleResults)

}
