% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printOccupancyCovariatesEffect}
\alias{printOccupancyCovariatesEffect}
\title{Estimates of covariates on occupancy probability}
\usage{
printOccupancyCovariatesEffect(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A data.frame with of the 95\% credible interval and the median.
}
\description{
Prints the estimates of the covariates on the occupancy probability.
}
\examples{

printOccupancyCovariatesEffect(sampleResults)

}
