% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{printOccupancyIndex}
\alias{printOccupancyIndex}
\title{Occupancy index}
\usage{
printOccupancyIndex(modelResults)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}
}
\value{
A data.frame with of the 95\% credible interval and the median.
}
\description{
Prints the estimates of the occupancy index, 
as described in Diana et al. (2021).
}
\examples{

printOccupancyIndex(sampleResults)

}
