% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodel.r
\name{runModel}
\alias{runModel}
\title{Model fitting}
\usage{
runModel(
  data,
  index_year,
  index_site,
  index_occ,
  index_spatial_x = 0,
  index_spatial_y = 0,
  covariates_psi_text,
  covariates_p_text,
  prior_psi = 0.5,
  sigma_psi = 2,
  prior_p = 0.5,
  sigma_p = 2,
  usingYearDetProb = F,
  usingSpatial = F,
  gridStep,
  storeRE = F,
  nchain,
  nburn,
  niter
)
}
\arguments{
\item{data}{The data frame containing the data.}

\item{index_year}{The index of the column containing the year variable.}

\item{index_site}{The index of the column containing the site variable.}

\item{index_occ}{The index of the column containing the detections.}

\item{index_spatial_x}{The index of the x coordinate of the site.}

\item{index_spatial_y}{The index of the y coordinate of the site.}

\item{covariates_psi_text}{Indexes of the column of the occupancy probability. To be separated by a comma, eg. "5,6,8". "0" is not available}

\item{covariates_p_text}{Indexes of the column of the detection probability.}

\item{prior_psi}{Prior mean for the occupancy probability}

\item{sigma_psi}{Standard deviation for the prior on the occupancy probability}

\item{prior_p}{Prior mean for the detection probability}

\item{sigma_p}{Standard deviation for the prior on the detection probability}

\item{usingYearDetProb}{Should the model include year-specific detection probabilities (as opposed to a constant one)?}

\item{usingSpatial}{Should the model include the auto-correlated spatial effects?}

\item{gridStep}{Step of the grid to use for the auto-correlated spatial effects. Use \code{buildSpatialGrid} to show the grid for a value of \code{gridStep}.}

\item{storeRE}{Should the model store the site-specific random effects for each iteration? Not suggested if the number of sites is greater than 1000.}

\item{nchain}{Number of chains.}

\item{nburn}{Number of burn-in iterations.}

\item{niter}{Number of (non burn-in) iterations.}
}
\value{
A list with components:

 \code{modelResults}: A list with components:
      
      \code{beta_psi_output}: An array of dimensions \code{nchain} x \code{niter} x (number of coefficients for 
      occupancy probability). The coefficients are reported in the order (year r.e., space r.e., covariates for time-space interaction,
      standard covariates) 
      
 \code{dataCharacteristics}: A list with components:
}
\description{
This function fits the occupancy model of Diana et al. (2021). 
To easily display the estimates of the quantities of interest, use the functions 
provided with the package.
}
\examples{

modelResults <- runModel(sampleData, 
                         index_year = 1, 
                         index_site = 2, 
                         index_occ = 8, 
                         index_spatial_x = 3, 
                         index_spatial_y = 4, 
                         covariates_psi_text = "5", 
                         covariates_p_text = "6-7", 
                         usingSpatial = TRUE,
                         gridStep = .2, 
                         nchain = 1, 
                         nburn = 100,
                         niter = 100)  

}
