% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_z_cpp}
\alias{sample_z_cpp}
\title{Sample the occupancies}
\usage{
sample_z_cpp(psi, p, k_s)
}
\arguments{
\item{psi}{Probabilities of detection.}

\item{k_s}{Correspondance.}
}
\description{
Sample the occupancies
}
