% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.r
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Sample datasets}
\format{
Sample dataset:
\describe{
  \item{Year}{Year when the observation is collected}
  \item{Site}{Site where the observation is collected}
  \item{X}{X coordinate of the site}
  \item{Y}{Y coordiante of the site}
  \item{OccCov1}{Sample covariate for the occupancy probability}
  \item{CaptureCovariate1}{Sample covariate for the detection probability}
  \item{CaptureCovariate1}{Sample covariate for the detection probability}
  \item{OCC}{Detection}
}
}
\usage{
sampleData
}
\description{
A dataset of 10 years and 500 sites
 with occupancy and detection covariates
}
\keyword{datasets}
