% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleResults.r
\docType{data}
\name{sampleResults}
\alias{sampleResults}
\title{Sample results}
\format{
An object of class \code{list} of length 2.
}
\usage{
sampleResults
}
\description{
An example of the output from the function \code{runModel}
}
\keyword{datasets}
