% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data function}
\usage{
simulateData(
  Y,
  S_years,
  V_lambda,
  mu_psi,
  b_t,
  beta_psi_tsp,
  beta_psi,
  X,
  a_s_site,
  mu_p,
  beta_p
)
}
\arguments{
\item{Y}{number of years}

\item{S_years}{number of sites per year}

\item{V_lambda}{}

\item{mu_psi}{}

\item{b_t}{}

\item{beta_psi_tsp}{}

\item{beta_psi}{}

\item{X}{}

\item{a_s_site}{}

\item{mu_p}{}

\item{beta_p}{}
}
\value{

}
\description{
Simulate data function
}
\examples{
simulateData(Y = 3, S_years = rep(50, 3),
             V_lambda = 2, mu_psi = 0, b_t = c(1,2,3), beta_psi_tsp = c(0,0),
             beta_psi = c(1), X = matrix(runif(50 * 2), 50, 2), a_s_site = rnorm(50),
              mu_p = -1, beta_p = 0)
}
