% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{tracePlot_DetectionCovariate}
\alias{tracePlot_DetectionCovariate}
\title{Traceplot of the covariates of the detection probability}
\usage{
tracePlot_DetectionCovariate(modelResults, index_cov)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}

\item{index_cov}{Index of the covariate to plot. Indexes go from 1 to the number of covariate.}
}
\value{
The traceplot with the estimate of the effective sample size.
}
\description{
Traceplot of the covariates coefficient, with 
effective sample size.
}
\examples{

tracePlot_DetectionCovariate(sampleResults, 1)

}
