% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{tracePlot_DetectionIntercept}
\alias{tracePlot_DetectionIntercept}
\title{Traceplot of the intercept of the detection probability}
\usage{
tracePlot_DetectionIntercept(modelResults, index_year)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}

\item{index_year}{Index of the year to plot, in case \code{usingYearDetProb} is 
set to \code{TRUE}. Indexes go from 1 to the number of years.}
}
\value{
The traceplot with the estimate of the effective sample size.
}
\description{
Traceplot of the intercept, with 
effective sample size.
}
\examples{

tracePlot_DetectionIntercept(sampleResults, 1)

}
