% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{tracePlot_OccupancyIndexRate}
\alias{tracePlot_OccupancyIndexRate}
\title{Traceplot of the occupancy index rate}
\usage{
tracePlot_OccupancyIndexRate(modelResults, index_year)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}

\item{index_year}{Index of the year to plot. Indexes go from 1 to the number of years.}
}
\value{
The traceplot with the estimate of the effective sample size.
}
\description{
Traceplot of the index occupancy rate, with 
effective sample size.
}
\examples{

tracePlot_OccupancyIndexRate(sampleResults, 1)

}
