% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{tracePlot_OccupancySiteEffect}
\alias{tracePlot_OccupancySiteEffect}
\title{Traceplot of the site-specific autocorrelated random effects}
\usage{
tracePlot_OccupancySiteEffect(modelResults, index_site)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}

\item{index_site}{Index of the random effect to plot. Indexes go from 1 to the number of sites in the approximation.}
}
\value{
The traceplot with the estimate of the effective sample size.
}
\description{
Traceplot of the site-specific random effects, with 
effective sample size.
}
\examples{

tracePlot_OccupancySiteEffect(sampleResults, 1)

}
