% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.r
\name{tracePlot_OccupancyYearEffect}
\alias{tracePlot_OccupancyYearEffect}
\title{Traceplot of the year-specific random effects}
\usage{
tracePlot_OccupancyYearEffect(modelResults, index_year)
}
\arguments{
\item{modelResults}{Output of the function \code{runModel}}

\item{index_year}{Index of the year of the random effect to plot. Indexes go from 1 to the number of years.}
}
\value{
The traceplot with the estimate of the effective sample size.
}
\description{
Traceplot of the year-specific random effects, with 
effective sample size..
}
\examples{

tracePlot_OccupancyYearEffect(sampleResults, 1)

}
