% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfun.r
\name{buildSpatialGrid}
\alias{buildSpatialGrid}
\title{Build spatial grid}
\usage{
buildSpatialGrid(X_cov, Y_cov, gridStep)
}
\arguments{
\item{X_cov}{The variable containing the x coordinates of the sites.}

\item{Y_cov}{The variable containing the y coordinates of the sites.}

\item{gridStep}{Value of the grid step.}
}
\value{
Displays of the grid with number of patches.
}
\description{
This function shows the spatial grid in the study area for a particular value of the grid step.
}
\examples{

buildSpatialGrid(sampleData$X, sampleData$Y, gridStep = .3)

}
